refactor(mitm-addon): inline precedence-only test helpers
